chore: refactor - extract common components from feature pages to shared
