+++ app/tag-data.json
@@ -17,7 +17,7 @@
   "events": 1,
   "python": 11,
   "log-management": 17,
-  "apm": 12,
+  "apm": 13,
   "security": 2,
   "claude-code": 1,
   "product-updates": 38,
@@ -26,7 +26,7 @@
   "cloud": 1,
   "aws": 3,
   "talks": 4,
-  "datadog": 3,
+  "datadog": 4,
   "pricing": 4,
   "observability": 58,
   "team": 1,
