+++ shared/components/molecules/FeaturePages/ButtonGroup/ButtonGroup.types.ts
@@ -1,0 +1,11 @@
+export interface ButtonGroupProps {
+    buttons: Array<{
+      text: string
+      href: string
+      variant: 'default' | 'secondary' | 'ghost'
+      icon?: React.ReactNode
+      className?: string
+    }>
+    className?: string
+  }
+
