+++ shared/components/molecules/FeaturePages/ButtonGroup/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './ButtonGroup.view'
