+++ shared/components/molecules/FeaturePages/FeatureCard/FeatureCard.types.ts
@@ -1,0 +1,8 @@
+export interface FeatureCardProps {
+    icon: React.ReactNode
+    title: string
+    description: string
+    variant?: 'default' | 'combined'
+    className?: string
+  }
+
