+++ shared/components/molecules/FeaturePages/FeatureCard/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './FeatureCard.view'
