+++ shared/components/molecules/FeaturePages/GridLayout/GridLayout.types.ts
@@ -1,0 +1,6 @@
+exprot interface GridLayoutProps {
+    children: React.ReactNode
+    cols?: number
+    className?: string
+    variant?: 'default' | 'equal' | 'split'
+  }
