+++ shared/components/molecules/FeaturePages/GridLayout/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './GridLayout.view'
