+++ shared/components/molecules/FeaturePages/HeroCards/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './HeroCards.view'
