+++ shared/components/molecules/FeaturePages/IconGrid/IconGrid.types.ts
@@ -1,0 +1,6 @@
+export interface IconGridProps {
+    icons: Array<{ src: string | React.ReactNode; alt: string }>
+    title: string
+    className?: string
+  }
+
