+++ shared/components/molecules/FeaturePages/IconGrid/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './IconGrid.view'
