+++ shared/components/molecules/FeaturePages/SectionLayout/SectionLayout.types.ts
@@ -1,0 +1,7 @@
+export interface SectionLayoutProps {
+    children: React.ReactNode
+    className?: string
+    variant?: 'default' | 'full-width' | 'bordered' | 'no-border' | 'border-x'
+    withBackground?: boolean
+  }
+
