+++ shared/components/molecules/FeaturePages/SectionLayout/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './SectionLayout.view'
+export type { SectionLayoutProps } from './SectionLayout.types'
