+++ shared/components/molecules/FeaturePages/SignozStats/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './SignozStats.view'
