+++ shared/components/molecules/FeaturePages/TestimonialCard/TestimonialCard.types.ts
@@ -1,0 +1,7 @@
+export interface TestimonialCardProps {
+    name: string
+    role: string
+    testimonial: string
+    image: string
+    className?: string
+  }
