+++ shared/components/molecules/FeaturePages/TestimonialCard/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './TestimonialCard.view'
