+++ shared/components/molecules/FeaturePages/UsageBasedPricing/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './UsageBasedPricing.view'
