+++ tsconfig.json
@@ -20,6 +20,9 @@
     "jsx": "preserve",
     "baseUrl": "./",
     "paths": {
+      "@/shared/*": [
+        "shared/*"
+      ],
       "@/components/*": [
         "components/*"
       ],
